I went through and fixed all the errors and warning for the html validator
I changed all the button containers to divs because I don't think that there is any better way to classify them. I also changed some of the p tag containers to divs, but I think there is a way to make it so that you don't have to use divs, but it gets rid of the warning on misused sections for now. What needs to be done is to format the rest of the page in the similar style and see if there is any way we can reduce the div count. There might not be just think about it and try to see if anything can be changed.