A file input page
I did all the js html and css needed to create a page to post content. Needs a header and footer to be put in when those are done, and also does not fully work because we don't have a backend, but should be pretty good in terms of set up for one.