login php
when user login, the rcsid(userID in the table) will be stored in the database, if the user is new to the web, it will create the new user in the table and make the username same with userID as default.
Also, please let me know what is in the users table, the admin col. Thanks.
** The function must be tested when uploaded to VM, sorry i have technical difficulty to do that.