Added dropdown options to the uploads page for tags and location
I added dropdown options to the uploads page for locations and tags this way they will be stored along with the other info. I also changed all the files to php files and used the same headers and footers for each one. ONE MASSIVE BUG IS ON THE MAIN PAGE I HAVE NO IDEA WHY THE FOOTER IS NOT AT THE BOTTOM