Backend and queries are set up sort of for the main page and the upload page
I converted almost all the pages to php. For the main page everything is now based off of queries and sessions. The user will be defaulted to a normal query where items show up by their postid ie time of post first. All the buttons on the side also when clicked on will be used as queries and will reprint what is already being printed out. Also for the uploads page, redid the user imported data. It now hopefully has functionality and the ability to put files into the database. For the main page I still need the comments to work as a modal and to create a function to increase the like counter by one when clicked on among other things.