2 main things were done 1st fixed bugs on the main page 2. added hash to upload
So first off I fixed some bugs on the main page so now there are no errors when printing things and also all the buttons and searching work as intended.
The next thing I did was I added hashing to the pictures and stuff. It should all work, but its hard to actually test locally so pls look it over well. I also added a bit of php to make sure that a png or jpg or jepg file was uploaded and only then can you actually upload to the database